requirements.txt
# Core AI & Agents
google-generativeai>=0.4.1
google-adk>=0.1.0

# Web & API
requests>=2.31.0
httpx>=0.27.0

# Data & File Processing
python-docx>=1.1.0
pypdf>=4.1.0
python-dotenv>=1.0.1
tomli>=2.0.1; python_version < "3.11"  # For Python < 3.11 compatibility
# tomllib is built-in for Python 3.11+
PyPDF2

# Utilities
pycountry>=23.12.11

# Voice Stack (Optional but recommended for full features)
# Note: These might have system dependencies like ffmpeg
SpeechRecognition>=3.10.1
edge-tts>=6.1.10
pygame>=2.5.2
faster-whisper>=0.10.1

protobuf>=5.29.5




